ISSUE #2418: Moved page requests to dedicated request handler
