+++ templates/html/index.html.twig
@@ -88,7 +88,7 @@
                     </a>
                 </li>
                 <li>
-                    <a title="Eddington" href="{{ relativeUrl('eddington') }}" data-router-navigate="{{ relativeUrl('eddington') }}" data-router-content-url="{{ relativeUrl('api/eddington') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
+                    <a title="Eddington" href="{{ relativeUrl('eddington') }}" data-router-navigate="{{ relativeUrl('eddington') }}" data-router-content-url="{{ relativeUrl('api/page/eddington') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
                         <svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
                             <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 4.5V19a1 1 0 0 0 1 1h15M7 14l4-4 4 4 5-5m0 0h-3.207M20 9v3.207"/>
                         </svg>
@@ -130,7 +130,7 @@
                 <ul class="space-y-1 py-2">
                     {% if completedChallenges > 0 %}
                         <li>
-                            <a title="{{ "Challenges"|trans }}" href="{{ relativeUrl('challenges') }}" data-router-navigate="{{ relativeUrl('challenges') }}" data-router-content-url="{{ relativeUrl('api/challenges') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
+                            <a title="{{ "Challenges"|trans }}" href="{{ relativeUrl('challenges') }}" data-router-navigate="{{ relativeUrl('challenges') }}" data-router-content-url="{{ relativeUrl('api/page/challenges') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
                                 <svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
                                     <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="m8.032 12 1.984 1.984 4.96-4.96m4.55 5.272.893-.893a1.984 1.984 0 0 0 0-2.806l-.893-.893a1.984 1.984 0 0 1-.581-1.403V7.04a1.984 1.984 0 0 0-1.984-1.984h-1.262a1.983 1.983 0 0 1-1.403-.581l-.893-.893a1.984 1.984 0 0 0-2.806 0l-.893.893a1.984 1.984 0 0 1-1.403.581H7.04A1.984 1.984 0 0 0 5.055 7.04v1.262c0 .527-.209 1.031-.581 1.403l-.893.893a1.984 1.984 0 0 0 0 2.806l.893.893c.372.372.581.876.581 1.403v1.262a1.984 1.984 0 0 0 1.984 1.984h1.262c.527 0 1.031.209 1.403.581l.893.893a1.984 1.984 0 0 0 2.806 0l.893-.893a1.985 1.985 0 0 1 1.403-.581h1.262a1.984 1.984 0 0 0 1.984-1.984V15.7c0-.527.209-1.031.581-1.403Z"/>
                                 </svg>
@@ -141,7 +141,7 @@
                     {% endif %}
                     {% if totalPhotoCount > 0 %}
                         <li>
-                            <a title="{{ "Photos"|trans }}" href="{{ relativeUrl('photos') }}" data-router-navigate="{{ relativeUrl('photos') }}" data-router-content-url="{{ relativeUrl('api/photos') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
+                            <a title="{{ "Photos"|trans }}" href="{{ relativeUrl('photos') }}" data-router-navigate="{{ relativeUrl('photos') }}" data-router-content-url="{{ relativeUrl('api/page/photos') }}" {{ isSymfonyRouter ? 'data-router-disabled' : '' }}>
                                 <svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
                                     <path stroke="currentColor" stroke-linejoin="round" stroke-width="2" d="M4 18V8a1 1 0 0 1 1-1h1.5l1.707-1.707A1 1 0 0 1 8.914 5h6.172a1 1 0 0 1 .707.293L17.5 7H19a1 1 0 0 1 1 1v10a1 1 0 0 1-1 1H5a1 1 0 0 1-1-1Z"/>
                                     <path stroke="currentColor" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z"/>
